Block VT switching for F12 via xkbcomp
This has the advantage of happening at runtime, without any need to
recompile.

Unfortunately this is at the same time a disadvantage, applying and or
patching the xkb configuration might fail silently if there are changes
in Xserver/Xkeyboard. Testing would have to be done for each keymap
(each language).