Create separate partition for EFI bootloader
This changes the installer to create a separate FAT32 partition that is
exclusively used for the EFI bootloader. Other boot and rescue system
related data is stored on an ext4 partition with journaling to
guard against data corruption.

The ext4 partition is mounted at `/boot` and the EFI system partition is
mounted underneath at `/boot/efi`.

In order to maintain backwards compatibility the following wart needs to
be introduced: The ext4 partition, which is _not_ the EFI system
partition needs to have the label `ESP`. The actual EFI system partition
can receive any other label.

This is necessary in order for both prior and following versions of
PlayOS to mount the right partition as `/boot` (PlayOS does not mount or
interact with the actual bootloader partition once booted).

The configuration in this commit has been tested to be backwards
compatible by installing a system using the previous installer (yielding
prior partition layout), then updating and booting one of its system
partitions via RAUC. Both system partitions remain operational and are
able to work with `/boot` (in this case the old FAT32 partition).