Treat complete but strange responses as captive
Add 511 as an explicitly expected result.

In addition, treat any other completed request with unexpected result as
indicative of a captive portal. Of course this means we may cause false
positives or temporary confusion if there is a problem with our probing
endpoints. But there is no knowing which dirty deeds some networks may
do. The classic case would be returning `200` with a completely
different body. But sysadmins or developers may have gotten creative and
return a 401[1].

[1]: https://knowledgebase.paloaltonetworks.com/KCSArticleDetail?id=kA10g000000ClJYCA0