Disable ConnMan's online check
Enabled back in https://github.com/dividat/playos/pull/18, mostly due to
a misunderstanding of which services the online check is applied to, and
therefore expecting it to help avoiding use of a Senso-connected
interface as default route in more cases than it actually could.

The online check however can cause ConnMan 1.42 to hang if an HTTPS
request is triggered, and also does not actually work with HTTPS in
later versions. This means none of the hosts we already ask customers to
whitelist is useful as a check URL, while false negative checks due to
firewall limitations might cause ConnMan to take unwarranted
"corrective" actions.