Apply a patch to make connman sorting sensitive to IPv4 link-local
This applies a patch to connman to make its service sorting factor in
the type of IPv4 method used by services. Link-local/Auto-IP configured
services are demoted relative to connected services using other methods.

This happens relatively late in the sorting hierarchy so that

- online services still win over ready services,
- ready services still win over failed-online-check services,
- preferred technologies still win over less preferred.

It may be wise to reconsider in detail which of the above should be
preserved.

This patch is a hacky PoC and seems to establish that the sorting logic
works, but additional hooks are required to make connman re-sort.